+++ src/intellisense/intellisenseProvider.node.ts
@@ -18,9 +18,9 @@ import { IDisposableRegistry, IConfigurationService, IsPreRelease } from '../pla
 import { IInterpreterService } from '../platform/interpreter/contracts';
 import { PythonEnvironment } from '../platform/pythonEnvironments/info';
 import { getInterpreterId } from '../platform/pythonEnvironments/info/interpreter';
-import { INotebookControllerManager, INotebookCompletionProvider, INotebookEditorProvider } from '../notebooks/types';
+import { INotebookCompletionProvider, INotebookEditorProvider } from '../notebooks/types';
 import { LanguageServer } from './languageServer.node';
-import { IVSCodeNotebookController } from '../notebooks/controllers/types';
+import { IControllerSelection, IVSCodeNotebookController } from '../notebooks/controllers/types';
 import { getComparisonKey } from '../platform/vscode-path/resources';
 import { CompletionRequest } from 'vscode-languageclient';
 import { NotebookPythonPathService } from './notebookPythonPathService.node';
@@ -43,7 +43,7 @@ export class IntellisenseProvider implements INotebookCompletionProvider, IExten
 
     constructor(
         @inject(IDisposableRegistry) private readonly disposables: IDisposableRegistry,
-        @inject(INotebookControllerManager) private readonly notebookControllerManager: INotebookControllerManager,
+        @inject(IControllerSelection) private readonly notebookControllerSelection: IControllerSelection,
         @inject(INotebookEditorProvider) private readonly notebookEditorProvider: INotebookEditorProvider,
         @inject(IVSCodeNotebook) private readonly notebooks: IVSCodeNotebook,
         @inject(IInterpreterService) private readonly interpreterService: IInterpreterService,
@@ -56,7 +56,7 @@ export class IntellisenseProvider implements INotebookCompletionProvider, IExten
 
     public activate() {
         // Sign up for kernel change events on notebooks
-        this.notebookControllerManager.onNotebookControllerSelected(this.controllerChanged, this, this.disposables);
+        this.notebookControllerSelection.onControllerSelected(this.controllerChanged, this, this.disposables);
         // Sign up for notebook open and close events.
         this.notebooks.onDidOpenNotebookDocument(this.openedNotebook, this, this.disposables);
         this.notebooks.onDidCloseNotebookDocument(this.closedNotebook, this, this.disposables);
@@ -74,7 +74,7 @@ export class IntellisenseProvider implements INotebookCompletionProvider, IExten
     }
 
     public async getLanguageClient(notebook: NotebookDocument) {
-        const controller = this.notebookControllerManager.getSelectedNotebookController(notebook);
+        const controller = this.notebookControllerSelection.getSelected(notebook);
         const interpreter = controller
             ? controller.connection.interpreter
             : await this.interpreterService.getActiveInterpreter(notebook.uri);
@@ -165,7 +165,7 @@ export class IntellisenseProvider implements INotebookCompletionProvider, IExten
             !this.notebookPythonPathService.isPylanceUsingLspNotebooks()
         ) {
             // Create a language server as soon as we open. Otherwise intellisense will wait until we run.
-            const controller = this.notebookControllerManager.getSelectedNotebookController(n);
+            const controller = this.notebookControllerSelection.getSelected(n);
 
             // Save mapping from notebook to controller
             if (controller) {
@@ -207,9 +207,7 @@ export class IntellisenseProvider implements INotebookCompletionProvider, IExten
         // We should allow intellisense for a URI when the interpreter matches
         // the controller for the uri
         const notebook = this.notebookEditorProvider.findAssociatedNotebookDocument(uri);
-        const controller = notebook
-            ? this.notebookControllerManager.getSelectedNotebookController(notebook)
-            : undefined;
+        const controller = notebook ? this.notebookControllerSelection.getSelected(notebook) : undefined;
         const notebookInterpreter = controller ? controller.connection.interpreter : this.getActiveInterpreterSync(uri);
         let notebookId = notebookInterpreter ? this.getInterpreterIdFromCache(notebookInterpreter) : undefined;
