+++ src/notebooks/languages/emptyNotebookCellLanguageService.ts
@@ -10,9 +10,8 @@ import { traceError } from '../../platform/logging';
 import { IDisposableRegistry } from '../../platform/common/types';
 import { noop } from '../../platform/common/utils/misc';
 import { chainWithPendingUpdates } from '../../kernels/execution/notebookUpdater';
-import { INotebookControllerManager } from '../types';
 import { isJupyterNotebook, translateKernelLanguageToMonaco } from '../../platform/common/utils';
-import { IVSCodeNotebookController } from '../controllers/types';
+import { IControllerSelection, IVSCodeNotebookController } from '../controllers/types';
 /**
  * If user creates a blank notebook, then they'll mostl likely end up with a blank cell with language, lets assume `Python`.
  * Now if the user changes the kernel to say `Julia`. After this, they need to also change the language of the cell.
@@ -25,10 +24,10 @@ export class EmptyNotebookCellLanguageService implements IExtensionSingleActivat
     constructor(
         @inject(IVSCodeNotebook) private readonly notebook: IVSCodeNotebook,
         @inject(IDisposableRegistry) private readonly disposables: IDisposableRegistry,
-        @inject(INotebookControllerManager) private readonly notebookControllerManager: INotebookControllerManager
+        @inject(IControllerSelection) private readonly notebookControllerSelection: IControllerSelection
     ) {}
     public async activate(): Promise<void> {
-        this.notebookControllerManager.onNotebookControllerSelected(
+        this.notebookControllerSelection.onControllerSelected(
             this.onDidChangeNotebookController,
             this,
             this.disposables
