+++ src/notebooks/types.ts
@@ -4,7 +4,6 @@ import {
     CancellationToken,
     CompletionContext,
     CompletionItem,
-    Event,
     NotebookDocument,
     NotebookEditor,
     Position,
@@ -12,45 +11,8 @@ import {
     Uri
 } from 'vscode';
 import { Resource } from '../platform/common/types';
-import { KernelConnectionMetadata, LiveRemoteKernelConnectionMetadata } from '../kernels/types';
-import { IVSCodeNotebookController } from './controllers/types';
-import { JupyterNotebookView, InteractiveWindowView } from '../platform/common/constants';
 
 export const INotebookKernelResolver = Symbol('INotebookKernelResolver');
-
-export const INotebookControllerManager = Symbol('INotebookControllerManager');
-export interface INotebookControllerManager {
-    readonly onNotebookControllerSelected: Event<{ notebook: NotebookDocument; controller: IVSCodeNotebookController }>;
-    readonly onNotebookControllerSelectionChanged: Event<{
-        notebook: NotebookDocument;
-        controller: IVSCodeNotebookController;
-        selected: boolean;
-    }>;
-    readonly kernelConnections: Promise<Readonly<KernelConnectionMetadata>[]>;
-    readonly remoteRefreshed: Event<LiveRemoteKernelConnectionMetadata[]>;
-    /**
-     * @param {boolean} [refresh] Optionally forces a refresh of all local/remote kernels.
-     */
-    loadNotebookControllers(refresh?: boolean): Promise<void>;
-    getSelectedNotebookController(document: NotebookDocument): IVSCodeNotebookController | undefined;
-    // Marked test only, just for tests to access registered controllers
-    getRegisteredNotebookControllers(): IVSCodeNotebookController[];
-    getActiveInterpreterOrDefaultController(
-        notebookType: typeof JupyterNotebookView | typeof InteractiveWindowView,
-        resource: Resource
-    ): Promise<IVSCodeNotebookController | undefined>;
-    getControllerForConnection(
-        connection: KernelConnectionMetadata,
-        notebookType: typeof JupyterNotebookView | typeof InteractiveWindowView
-    ): IVSCodeNotebookController | undefined;
-    getPreferredNotebookController(document: NotebookDocument): IVSCodeNotebookController | undefined;
-    initializePreferredNotebookController(document: NotebookDocument): Promise<void>;
-    computePreferredNotebookController(
-        document: NotebookDocument,
-        serverId?: string
-    ): Promise<{ preferredConnection?: KernelConnectionMetadata; controller?: IVSCodeNotebookController }>;
-}
-
 export const INotebookCompletionProvider = Symbol('INotebookCompletionProvider');
 
 export interface INotebookCompletionProvider {
