+++ src/webviews/extension-side/api/api.ts
@@ -12,7 +12,7 @@ import {
     IJupyterUriProviderRegistration,
     JupyterServerUriHandle
 } from '../../../kernels/jupyter/types';
-import { INotebookControllerManager, INotebookEditorProvider } from '../../../notebooks/types';
+import { INotebookEditorProvider } from '../../../notebooks/types';
 import { IDataViewerDataProvider, IDataViewerFactory } from '../dataviewer/types';
 import { IExportedKernelService } from './extension';
 import { IPythonApiProvider, PythonApi } from '../../../platform/api/types';
@@ -20,6 +20,7 @@ import { isTestExecution } from '../../../platform/common/constants';
 import { IExtensionContext } from '../../../platform/common/types';
 import { IServiceContainer, IServiceManager } from '../../../platform/ioc/types';
 import { traceError } from '../../../platform/logging';
+import { IControllerPreferredService } from '../../../notebooks/controllers/types';
 
 export const IExportedKernelServiceFactory = Symbol('IExportedKernelServiceFactory');
 export interface IExportedKernelServiceFactory {
@@ -120,7 +121,7 @@ export function buildApi(
             handle: JupyterServerUriHandle,
             notebook: NotebookDocument
         ) => {
-            const controllers = serviceContainer.get<INotebookControllerManager>(INotebookControllerManager);
+            const controllers = serviceContainer.get<IControllerPreferredService>(IControllerPreferredService);
             const connection = serviceContainer.get<JupyterConnection>(JupyterConnection);
             const uri = generateUriFromRemoteProvider(providerId, handle);
             await connection.updateServerUri(uri);
